Apply crispy layout to base DRF FilterSet
